Module spin_sync::mutex

Structs

Mutex
A mutual exclusion primitive useful for protecting shared data

MutexGuard
An RAII implementation of a "scoped lock" of a mutex.

PoisonError
A type of error which can be returned whenever a lock is acquired.

Enums

TryLockError
An enumeration of possible errors associated with a TryLockResult which can occur while trying to acquire a lock, from the try_lock method on a Mutex or the try_read and try_write methods on an RwLock.

Type Definitions

LockResult
A type alias for the result of a lock method which can be poisoned.

TryLockResult
A type alias for the result of a nonblocking locking method.